stok icon dynamicaly added

+++ src/Hooks/UseProductDelete.js
@@ -1,12 +1,15 @@
 import axios from 'axios';
+import { useAuthState } from 'react-firebase-hooks/auth';
 import Swal from 'sweetalert2';
+import auth from '../Firebase/Firebase.config';
 import UseNotify from './UseNotify';
 
 const UseProductDelete = (products, setProducts, changeState, setChangeState) => {
-    const { notifySuccess } = UseNotify();
+    const [user, ,] = useAuthState(auth);
+    const { notifySuccess, notifyError } = UseNotify();
 
     const handleDelete = async productId => {
-        const url = `http://localhost:5000/delete-product/${productId}`;
+        const url = `https://protected-waters-02155.herokuapp.com/delete-product/${productId}`;
         try {
             Swal.fire({
                 title: "Are your sure?",
@@ -25,6 +28,10 @@ const UseProductDelete = (products, setProducts, changeState, setChangeState) =>
             })
                 .then(async (result) => {
                     if (result.isConfirmed) {
+                        if (user.email !== 'rumanislam0429@gmail.com') {
+                            notifyError('You are not authorized to delete.');
+                            return
+                        }
                         const { data } = await axios.delete(url)
                         if (data.success) {
                             const remainingProducts = products.filter(product => product._id !== productId);
